+++ salesforce/dbapi/driver.py
@@ -135,7 +135,8 @@ class RawConnection:
         pass
 
     def rollback(self) -> None:  # pylint:disable=no-self-use
-        log.info("Rollback is not implemented in Salesforce.")
+        if self.settings_dict.get('OPTIONS', {}).get('WARNING_ON_ROLLBACK', False):
+            log.info("Rollback is not implemented in Salesforce.")
 
     @overload
     def cursor(self) -> 'Cursor[Tuple[Any, ...]]': ...
@@ -513,7 +514,7 @@ class Cursor(Generic[_TRow]):
     for complete rows, one by one.
     """
 
-    # pylint:disable=too-many-instance-attributes
+    # pylint:disable=too-many-instance-attributes,too-many-public-methods
     def __init__(self, connection: Connection, row_type: Optional[Type[_TRow]] = None) -> None:
         # DB API attributes (public, ordered by documentation PEP 249)
         self.description = None           # type: Optional[List[Tuple[Any, ...]]]
@@ -620,6 +621,12 @@ class Cursor(Generic[_TRow]):
 
     def __next__(self) -> _TRow:
         return self._iter.__next__()
+
+    def commit(self) -> None:
+        self._connection.commit()
+
+    def rollback(self) -> None:
+        self._connection.rollback()
 
     # -- private methods